Ujednolicenie pakietyzacji
