+++ src/main/java/ja/workshops/spring/części/konfiguracja/xml/b_scope_i_import/package-info.java
@@ -11,9 +11,9 @@
  *  <h2> 1. Kilka plików konfiguracyjnych</h2>
  *      <p>W pakiecie dostępne są klasy z poprzedniego zadania. <font color="yellow">W katalogu resources w pakiecie b </font> znajdują się pliki konfiguracyjne xml.</p>
  *      <p>Jak można zauważyć, każde ziarno posiada swój własny plik konfiguracyjny.</p>
- *      <p>Zastanów się w jaki sposób możemy w takim wypadku wstrzykiwać zależności do klasy {@link ja.workshop.spring.części.konfiguracja.xml.b_scope_i_import.Samochód}?</p>
+ *      <p>Zastanów się w jaki sposób możemy w takim wypadku wstrzykiwać zależności do klasy {@link ja.workshops.spring.części.konfiguracja.xml.b_scope_i_import.Samochód}?</p>
  *
- *      <p>Spróbuj odpalić klasę {@link ja.workshop.spring.części.konfiguracja.xml.b_scope_i_import.Main} a następnie napraw występujące problemy.</p>
+ *      <p>Spróbuj odpalić klasę {@link ja.workshops.spring.części.konfiguracja.xml.b_scope_i_import.Main} a następnie napraw występujące problemy.</p>
  *      <p><b>Podpowiedź:</b> modyfikacja pliku <font color="yellow">{@code b_samochod_konfiguracja.xml}</font> i skorzystanie z {@code import}.</p>
  *  <br>
  *  <h2> 2. Zasięg ziarna</h2>
@@ -25,4 +25,4 @@
  * @author Wiktor Rup
  *
  */
-package ja.workshop.spring.części.konfiguracja.xml.b_scope_i_import;
+package ja.workshops.spring.części.konfiguracja.xml.b_scope_i_import;
