+++ src/main/java/ja/workshops/spring/części/konfiguracja/xml/b_scope_i_import/silnik/OlejNapędowy.java
@@ -1,0 +1,15 @@
+package ja.workshops.spring.części.konfiguracja.xml.b_scope_i_import.silnik;
+
+import ja.workshops.spring.części.konfiguracja.xml.b_scope_i_import.Samochód;
+
+/**
+ * Ziarno spingowe wstrzykiwane do klasy {@link Samochód Samochód}.
+ *
+ * @author Wiktor Rup
+ */
+public class OlejNapędowy implements Silnik {
+
+    public String start() {
+        return "Wrumm!";
+    }
+}
