+++ src/main/java/ja/workshops/spring/części/konfiguracja/xml/b_scope_i_import/silnik/Silnik.java
@@ -1,4 +1,4 @@
-package ja.workshop.spring.części.konfiguracja.xml.b_scope_i_import.silnik;
+package ja.workshops.spring.części.konfiguracja.xml.b_scope_i_import.silnik;
 
 /**
  * Interfejs na ziarna.
