+++ src/main/java/ja/workshops/spring/części/konfiguracja/xml/c_mvc/SerwisSamochodowy.java
@@ -1,4 +1,4 @@
-package ja.workshop.spring.części.konfiguracja.xml.c_mvc;
+package ja.workshops.spring.części.konfiguracja.xml.c_mvc;
 
 /**
  * Serwis, który jest warstwą łączącą dane wyciągane z repozytorium {@link RepozytoriumSamochodów} oraz warstwą prezentacji {@link KomisKontroler}.
